backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.13
pydantic-settings==2.4.0
python-dotenv==1.0.1